Telegram Group Reader/requirements.txt
# 🚀 Telegram Chat Manager - Production Requirements

# ============================================
# 🌐 Web Framework & Core
# ============================================
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
python-dotenv==1.0.1
werkzeug==3.0.3

# ============================================
# 🗄️ Database
# ============================================
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23 # Se si usa Flask-SQLAlchemy
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# ============================================
# 🔴 Redis & Cache
# ============================================
redis==5.0.7
hiredis==2.2.3

# ============================================
# 🤖 Telegram Integration
# ============================================
Telethon==1.34.0

# ============================================
# 🔐 Security & Authentication
# ============================================
Flask-Bcrypt==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0
cryptography==42.0.8
Flask-JWT-Extended==4.6.0

# ============================================
# 🔧 Utilities
# ============================================
requests==2.31.0
Pillow==10.1.0
blinker==1.7.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
click==8.1.7
colorama==0.4.6
docker==7.1.0

# ============================================
# 📊 Logging & Monitoring
# ============================================
structlog==23.2.0

# ============================================
# 📊 Validation & Forms
# ============================================
email-validator==2.1.0
phonenumbers==8.13.25

# ============================================
# 🌐 HTTP & Networking
# ============================================
urllib3==2.1.0
certifi==2023.11.17

# ============================================
# 📝 Logging & Monitoring
# ============================================
structlog==23.2.0 